refactor: move errorHandler middleware
